refactor: migration of metrics-slice.ts to zustand
