+++ frontend/src/store.ts
@@ -8,7 +8,6 @@ import commandReducer from "./state/command-slice";
 import { jupyterReducer } from "./state/jupyter-slice";
 import securityAnalyzerReducer from "./state/security-analyzer-slice";
 import statusReducer from "./state/status-slice";
-import metricsReducer from "./state/metrics-slice";
 import microagentManagementReducer from "./state/microagent-management-slice";
 import conversationReducer from "./state/conversation-slice";
 import eventMessageReducer from "./state/event-message-slice";
@@ -23,7 +22,6 @@ export const rootReducer = combineReducers({
   jupyter: jupyterReducer,
   securityAnalyzer: securityAnalyzerReducer,
   status: statusReducer,
-  metrics: metricsReducer,
   microagentManagement: microagentManagementReducer,
   conversation: conversationReducer,
   eventMessage: eventMessageReducer,
